test: relocate unit test cli, proxy, utils
